Remove support for mail gems before 2.7
Older versions of mail assumes that all values in Mail::Field objects are
Strings. We need to pass richer objects, usually Hashes, that we can forward to
the SendGrid API. For example when including variables for dynamic templates:

    mail(
      :to => "someone@example.com",
      :template_id => "d-42",
      :dynamic_variables => {
        :name => "Some One"
      }
    )

Previously, Mail::Field would convert the Hash

    {:dynamic_variables => {:name => "Some One"}}

to its String representation

    "{:dynamic_variables=>{:name=>\"Some One\"}}"

which we would then attempt to convert into something resembling JSON, so we
could parse it back into a Hash with the original values. Unfortunately this
would fail in cases where the value of the variable contains special characters,
in particular `=>`.